refactor: move token status list functionality to issuance domain
