+++ agent_issuance/src/credential/mod.rs
@@ -1,4 +1,5 @@
 pub mod aggregate;
+pub mod application;
 pub mod command;
 pub mod entity;
 pub mod error;
